Switch current route URLs
The changes are:
- root page (comparator) has been moved to /comparator
- /about has been moved to root page
