+++ src/pages/comparator.tsx
@@ -1,0 +1,16 @@
+import { NextSeo } from 'next-seo'
+
+import Layout from '~/components/Layout'
+import RepositoryReleasesComparator from '~/components/RepositoryReleasesComparator'
+import { ComparatorProvider } from '~/contexts/comparator-context'
+
+const IndexPage = () => (
+  <Layout>
+    <NextSeo title="Comparator" />
+    <ComparatorProvider>
+      <RepositoryReleasesComparator />
+    </ComparatorProvider>
+  </Layout>
+)
+
+export default IndexPage
